backend-fastapi/labs/D-data-api-lab/fastapi/pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "d-data-api-lab-fastapi"
version = "0.1.0"
description = "FastAPI lab for projects, tasks, comments, and optimistic locking."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "redis>=5.2,<6.0",
  "SQLAlchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "ruff>=0.8,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
